requirements.txt
flask==3.0.0
openai
Pillow
flask-limiter==3.5.0
flask-cors==4.0.0
python-dotenv==1.0.0 